# web/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_create_import_api_imports_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_create_import_api_imports_post",
        "type": "object"
      },
      "Body_create_import_api_v1_imports_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_create_import_api_v1_imports_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportAccepted": {
        "description": "The response to an upload: what was stored, what was not, and what happens next.\n\n`job_id` and `analysis_status` describe the analysis this upload queued. For an identical\nre-upload they are null - nothing new was queued - and `duplicate_upload` is true.",
        "properties": {
          "analysis_status": {
            "anyOf": [
              {
                "enum": [
                  "queued",
                  "running",
                  "succeeded",
                  "failed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Status"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "duplicate_upload": {
            "default": false,
            "title": "Duplicate Upload",
            "type": "boolean"
          },
          "duplicates_in_database": {
            "default": 0,
            "title": "Duplicates In Database",
            "type": "integer"
          },
          "duplicates_in_file": {
            "default": 0,
            "title": "Duplicates In File",
            "type": "integer"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "import_id": {
            "title": "Import Id",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "organisation_id": {
            "title": "Organisation Id",
            "type": "string"
          },
          "row_errors": {
            "items": {
              "$ref": "#/components/schemas/ImportRowError"
            },
            "title": "Row Errors",
            "type": "array"
          },
          "rows_imported": {
            "title": "Rows Imported",
            "type": "integer"
          },
          "rows_received": {
            "title": "Rows Received",
            "type": "integer"
          },
          "rows_rejected": {
            "title": "Rows Rejected",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "import_id",
          "organisation_id",
          "status",
          "rows_received",
          "rows_imported",
          "rows_rejected"
        ],
        "title": "ImportAccepted",
        "type": "object"
      },
      "ImportRowError": {
        "description": "One rejected row, identified by the line number in the customer's own file.",
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "row": {
            "title": "Row",
            "type": "integer"
          }
        },
        "required": [
          "row",
          "field",
          "message"
        ],
        "title": "ImportRowError",
        "type": "object"
      },
      "ImportSummary": {
        "description": "An import batch as a customer sees it.",
        "properties": {
          "analysis_status": {
            "anyOf": [
              {
                "enum": [
                  "queued",
                  "running",
                  "succeeded",
                  "failed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Status"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "import_id": {
            "title": "Import Id",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "organisation_id": {
            "title": "Organisation Id",
            "type": "string"
          },
          "rows_imported": {
            "title": "Rows Imported",
            "type": "integer"
          },
          "rows_received": {
            "title": "Rows Received",
            "type": "integer"
          },
          "rows_rejected": {
            "title": "Rows Rejected",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "import_id",
          "organisation_id",
          "status",
          "rows_received",
          "rows_imported",
          "rows_rejected"
        ],
        "title": "ImportSummary",
        "type": "object"
      },
      "JobSummary": {
        "description": "Progress of background work. `error` is a summary - never internal detail.",
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "max_attempts": {
            "title": "Max Attempts",
            "type": "integer"
          },
          "organisation_id": {
            "title": "Organisation Id",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "organisation_id",
          "kind",
          "status",
          "attempts",
          "max_attempts"
        ],
        "title": "JobSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "description": "Set by POST /api/v1/auth/login or /register. HttpOnly: a browser sends it automatically.",
        "in": "cookie",
        "name": "feedbackiq_session",
        "type": "apiKey"
      },
      "APIKeyHeader": {
        "in": "header",
        "name": "x-api-key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Backend for FeedbackIQ \u2014 sentiment analysis, complaint categorisation, semantic search and retrieval-augmented question answering over customer feedback.",
    "title": "FeedbackIQ API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/imports": {
      "post": {
        "description": "Validates the file, stores the valid rows for the signed-in user's organisation and queues analysis. Requires a `text` column; `external_id`, `created_at`, `rating` and `platform` are optional. Rejected rows are reported per row and do not stop the import. Re-uploading an identical file creates nothing and returns the original import.",
        "operationId": "create_import_api_imports_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_import_api_imports_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Upload a CSV of customer feedback",
        "tags": [
          "Imports"
        ]
      }
    },
    "/api/imports/{import_id}": {
      "get": {
        "operationId": "read_import_api_imports__import_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "title": "Import Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Status and counts for one import",
        "tags": [
          "Imports"
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "operationId": "read_job_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Status of a background job",
        "tags": [
          "Imports"
        ]
      }
    },
...
    "/api/v1/imports": {
      "get": {
        "description": "This organisation's imports, newest first, with the counts each produced.",
        "operationId": "list_imports_api_v1_imports_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ImportSummary"
                  },
                  "title": "Response List Imports Api V1 Imports Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Recent imports",
        "tags": [
          "Imports (v1)"
        ]
      },
      "post": {
        "description": "Validates the file, stores the valid rows for the signed-in user's organisation and queues analysis. Requires a `text` column; `external_id`, `created_at`, `rating` and `platform` are optional. Rejected rows are reported per row and do not stop the import. Re-uploading an identical file creates nothing and returns the original import.",
        "operationId": "create_import_api_v1_imports_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_import_api_v1_imports_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Upload a CSV of customer feedback",
        "tags": [
          "Imports (v1)"
        ]
      }
    },
    "/api/v1/imports/{import_id}": {
      "get": {
        "operationId": "read_import_api_v1_imports__import_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "title": "Import Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "One import",
        "tags": [
          "Imports (v1)"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "description": "Status, attempts and the analysis result summary. `error` is a short message, never internal detail.",
        "operationId": "read_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "One background job",
        "tags": [
          "Imports (v1)"
        ]
      }
    }
  }
}